# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckpointOut": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "integer"
          },
          "details": {
            "title": "Details",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sourceConversationId": {
            "title": "Sourceconversationid",
            "type": "string"
          },
          "sourceMessageId": {
            "title": "Sourcemessageid",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "details",
          "completed",
          "sourceConversationId",
          "sourceMessageId",
          "createdAt"
        ],
        "title": "CheckpointOut",
        "type": "object"
      },
      "CheckpointUpdate": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          }
        },
        "required": [
          "completed"
        ],
        "title": "CheckpointUpdate",
        "type": "object"
      },
      "CheckpointsResponse": {
        "properties": {
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointOut"
            },
            "title": "Checkpoints",
            "type": "array"
          }
        },
        "required": [
          "checkpoints"
        ],
        "title": "CheckpointsResponse",
        "type": "object"
      },
...
      "CreateNoteResponse": {
        "properties": {
          "note": {
            "$ref": "#/components/schemas/NoteOut"
          }
        },
        "required": [
          "note"
        ],
        "title": "CreateNoteResponse",
        "type": "object"
      },
...
      "DeleteNoteResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "DeleteNoteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NoteCreate": {
        "additionalProperties": false,
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Tags",
            "type": "array"
          },
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "NoteCreate",
        "type": "object"
      },
      "NoteOut": {
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "required": [
          "id",
          "text",
          "createdAt"
        ],
        "title": "NoteOut",
        "type": "object"
      },
      "NoteUpdate": {
        "additionalProperties": false,
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Tags",
            "type": "array"
          },
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "NoteUpdate",
        "type": "object"
      },
      "NotesResponse": {
        "properties": {
          "hasMore": {
            "default": false,
            "title": "Hasmore",
            "type": "boolean"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteOut"
            },
            "title": "Notes",
            "type": "array"
          },
          "searchLimited": {
            "default": false,
            "title": "Searchlimited",
            "type": "boolean"
          }
        },
        "required": [
          "notes"
        ],
        "title": "NotesResponse",
        "type": "object"
      },
      "UpdateCheckpointResponse": {
        "properties": {
          "checkpoint": {
            "$ref": "#/components/schemas/CheckpointOut"
          }
        },
        "required": [
          "checkpoint"
        ],
        "title": "UpdateCheckpointResponse",
        "type": "object"
      },
      "UpdateNoteResponse": {
        "properties": {
          "note": {
            "$ref": "#/components/schemas/NoteOut"
          }
        },
        "required": [
          "note"
        ],
        "title": "UpdateNoteResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "FirebaseBearer": {
        "description": "Firebase ID token issued to the signed-in NoteVault user.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NoteVault API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/checkpoints": {
      "get": {
        "operationId": "list_checkpoints_checkpoints_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Checkpoints",
        "tags": [
          "checkpoints"
        ]
      }
    },
    "/checkpoints/{checkpoint_id}": {
      "patch": {
        "operationId": "update_checkpoint_checkpoints__checkpoint_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "checkpoint_id",
            "required": true,
            "schema": {
              "title": "Checkpoint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCheckpointResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Update Checkpoint",
        "tags": [
          "checkpoints"
        ]
      }
    },
...
    "/notes": {
      "get": {
        "operationId": "list_notes_notes_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Notes",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "create_note_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Create Note",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes/{note_id}": {
      "delete": {
        "operationId": "delete_note_notes__note_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteNoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Delete Note",
        "tags": [
          "notes"
        ]
      },
      "patch": {
        "operationId": "update_note_notes__note_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateNoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Update Note",
        "tags": [
          "notes"
        ]
      }
    }
  }
}